feat: conexão do codastro de usuario com a api
