+++ sistemadetroca-front/app/page.js
@@ -1,34 +1,8 @@
-"use client";
-import { useState } from 'react';
+
 import Image from "next/image";
+import HendleSignup from "components/handleSignup"
 
 export default function Login() {
-  const [username, setUsername] = useState('');
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [birthdate, setBirthdate] = useState('');
-
-  const handleSignup = async () => {
-    try {
-      const response = await fetch('url_do_seu_backend/signup', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ username, email, password, birthdate }),
-      });
-
-      if (response.ok) {
-        // Lógica para sucesso na criação de conta
-        console.log('Conta criada com sucesso!');
-      } else {
-        // Lógica para tratamento de erro
-        console.error('Falha ao criar conta:', response.statusText);
-      }
-    } catch (error) {
-      console.error('Erro ao processar criação de conta:', error);
-    }
-  };
 
   return (
     <main className="flex h-dvh justify-center items-center">
@@ -55,58 +29,7 @@ export default function Login() {
           </div>
         </div>
 
-        <div className="w-[50rem] rounded-e-2xl">
-          <div className="flex flex-col justify-center items-center h-[30rem] gap-4">
-            <div className="flex flex-col justify-center items-center h-[10rem]">
-              <span className="text-[#F26329] font-bold text-[35px]">Crie uma conta</span>
-              <span className="text-slate-400">Insira as informações solicitadas abaixo</span>
-            </div>
-            <div className="flex w-[50%] p-3 bg-slate-100 gap-1">
-              {/* Ícone de usuário */}
-              <input
-                value={username}
-                onChange={(e) => setUsername(e.target.value)}
-                className="w-[90%] bg-slate-100 focus:outline-none"
-                placeholder="Usuário"
-              />
-            </div>
-            <div className="flex w-[50%] p-3 bg-slate-100 gap-1">
-              {/* Ícone de email */}
-              <input
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                type="email"
-                className="w-[90%] bg-slate-100 focus:outline-none"
-                placeholder="Email"
-              />
-            </div>
-            <div className="flex w-[50%] p-3 bg-slate-100 gap-1">
-              {/* Ícone de senha */}
-              <input
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                type="password"
-                className="w-[90%] bg-slate-100 focus:outline-none"
-                placeholder="Senha"
-              />
-            </div>
-            <div className="flex w-[50%] p-3 bg-slate-100 gap-1">
-              {/* Ícone de calendário */}
-              <input
-                value={birthdate}
-                onChange={(e) => setBirthdate(e.target.value)}
-                type="date"
-                className="w-[90%] bg-slate-100 focus:outline-none"
-              />
-            </div>
-            <div className="flex w-[50%] justify-center">
-              {/* Botão de criar conta */}
-              <button onClick={handleSignup} className="bg-[#F26329] p-2 rounded-lg text-white">
-                Criar Conta
-              </button>
-            </div>
-          </div>
-        </div>
+        <HendleSignup/>
 
       </div>
     </main>
